Move DD algorith plugins in plugin directory
