+++ DetectorDescription/Algorithm/plugins/DDAngular.h
